working: tab correctly jumps to header on scroll
emphasis on the 'on scroll'. If the header is already in the page and there is no possible way to scroll (e.g. the page is scrolled to the complete bottom, but there is another h1 to jump to), the tab won't move. The page will 'navigate' but since there is no scroll, there is no scroll event firing. I need to wait until the page finishes scrolling before it happens. If there is no scroll, perform the behavior. Maybe I need to know if the page is scrolling?